Simplify register_ranges method

+++ include/fuzzy/agenda_item.hh
@@ -18,6 +18,6 @@ namespace fuzzy
     std::vector<bool> map_pattern; /* set of unigrams that are both in the pattern and in the suffix
                                       map_pattern[i]==true means that the word pattern[i] appears in the suffix */
     int coverage; // number of true values in map_pattern
-    int maxmatch; // maximum length of the longest subsequence between the sentence associated with the AgendaItem and the pattern
+    size_t maxmatch; // maximum length of the longest subsequence between the sentence associated with the AgendaItem and the pattern
   };
 }
